feat(html-loading): Register HTML template with HTML service on require
We want to move toward a late binding approach for HTML templates, where we load up
the file content but don't translate, append, or instantiate the template until a
template request is made. This fixes ordering bugs around requiring translations
after templates.

This also means there is less code required to use a template, simply require
it and the loader will handle the registering of it with the HTML service.

closes: CTNXT-478