fix(brjs-converter): Post conversion script can modify `srcPath`
When updating the module paths following a conversion the code being analyzed may not be in the same location as it will eventually be following a build.
Given that the relative path that is calculated can end up being incorrect so we need to allow users to modify the `srcPath` value to its notional final value.

closes: CTNXT-501