feat(build-tasks): Change meaning of `warName` parameter
The `warName` parameter is no longer prefixed to the `version` to calculate
the name of the war file. Instead the name is used as provided, this allows
for more flexible naming. Required by ST which has variants of the application
and the variant name is appended to the end of the WAR name.

Breaking change.