feat(karma-config): Allow packages to specify plugins
Packages might have their own versions of Karma plugins which
clash with the more commonly used ones which because they are
more common are installed in the app's `node_modules` due to
package hoisting but the less common one would be installed
inside the package's `node_modules` e.g. the older version of
karma-jasmine vs the latest one.

This means we need to allow packages to specify the exact
`karma-jasmine` plugin to use, which can be done via the
`plugins` config array, but allowing the packages Karma config
to create an empty array with only the plugin they want to
configure in results in the default `plugin` value being
dropped so we need to copy the default value into our Karma
configuration.

Otherwise `karma-webpack` etc aren't loaded.