feat(proxy-factory): Terminate the server on proxy process errors
This prevents errors from being missed e.g. when running the development
stack on a new machine with no `mvn` installed, the Webpack/HTTP server
would work but the proxy target would fail to start.

While output indicating an error occured is printed out it may be missed
if there is a lot of other output.

BREAKING CHANGE: It's possible that certain errors don't affect the
running of the proxy target.