feat: Run Jasmine tests
Some Caplin test packages have a mixture of legacy JSTestDriver
tests and Jasmine tests. In that case two Karma adapters may be
loaded. When that happens we want to run both sets of tests so
we check for the existance of jasmine and if it exists we run
the previous Karma `start` function which would be the Jasmine
adapter `start` function.